+++ puppeteerServer/mangaGetProc.ts
@@ -6,12 +6,22 @@ import { getManga as getAsura } from './puppeteerScripts/asuraV3';
 import { getManga as getMangadex } from './puppeteerScripts/mangadex';
 import { getManga as getMangaPark } from './puppeteerScripts/mangapark';
 import { getManga as getMangafire } from './puppeteerScripts/mangafire';
+import { getManga as GetBato } from './puppeteerScripts/bato';
 
 export default async function (job: Job) {
   if (config.logging.verboseLogging) console.log('starting job \n' + job.data);
   switch (job.data.type) {
     case 'manganato':
       return await getManganato(
+        job.data.url,
+        job.data.getIcon,
+        job.data.update,
+        job.data.coverIndexes,
+        job.data.maxSavedAt,
+        job
+      );
+    case 'bato':
+      return await GetBato(
         job.data.url,
         job.data.getIcon,
         job.data.update,