+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # **Change Logs**
 
 ## Change Log 3.1.8
+ - Updated all puppeteer server packages.
+ - config option for browser path
+ - Updated queue viewer packages
+ - Updated all website Server packages
+ - Updated Some Client Page Packages
+ - Fixed mangapark base url
+ - Styled Change User Title Text Box
+ - Support for Bato
+
+## Change Log 3.1.8
  - MangaPark Support
 
 ## Change Log 2.0.3